+++ libs/core/src/state/settings/by_widget.rs
@@ -27,6 +27,19 @@ pub struct SettingsByWidget {
     pub others: HashMap<WidgetId, ThirdPartyWidgetSettings>,
 }
 
+impl SettingsByWidget {
+    pub fn is_enabled(&self, widget_id: &WidgetId) -> bool {
+        match widget_id.as_str() {
+            "@seelen/weg" => self.weg.enabled,
+            "@seelen/fancy-toolbar" => self.fancy_toolbar.enabled,
+            "@seelen/window-manager" => self.wm.enabled,
+            "@seelen/wallpaper-manager" => self.wall.enabled,
+            "@seelen/launcher" => self.launcher.enabled,
+            _ => self.others.get(widget_id).is_some_and(|s| s.enabled),
+        }
+    }
+}
+
 #[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema, TS)]
 #[serde(default)]
 pub struct ThirdPartyWidgetSettings {